+++ modules.d.ts
@@ -1,0 +1,14 @@
+declare module '*.wgsl' {
+  const content: string
+  export default content
+}
+
+declare module '*.svg' {
+  const content: string
+  export default content
+}
+
+declare module '*.ttf' {
+  const content: string
+  export default content
+}
