+++ src/utils/getAngle.ts
@@ -1,3 +1,5 @@
+import { Point } from 'types'
+
 export default function getAngle(from: Point, to: Point) {
   return Math.atan2(to.y - from.y, to.x - from.x)
 }
