+++ src/utils/normalizeVec2.ts
@@ -1,3 +1,5 @@
+import { Point } from 'types'
+
 export default function normalizeVec2(vec: Point): Point {
   const length = Math.sqrt(vec.x * vec.x + vec.y * vec.y)
   return {
